Moved language view helpers to sperate module
